+++ lib/moussa_hls_player.dart
@@ -1,4 +1,3 @@
-
 export 'src/moussa_hls_player_view.dart';
 export 'src/moussa_hls_player_controller.dart';
 export 'src/moussa_hls_player_types.dart';